# doc/api.yaml
openapi: 3.0.3
info:
  title: WASATEXT
  version: "1.0.0"
  description: |-
    API for the WASAText messaging application
  contact:
    name: Giordana Foglia
    email: foglia.2046738@studenti.uniroma1.it
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ConversationId:
      name: conversation_id
      in: path
      required: true
      description: "The unique identifier of the conversation"
      schema:
        $ref: "#/components/schemas/ConversationId"
...
  schemas:
...
    ConversationId:
      type: integer
      description: "Unique identifier for the conversation"
      minimum: 0
      maximum: 9999
      readOnly: true
      example: 823
...
    Name:
      type: string
      description: "username for an user"
      example: "Matteo"
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
...
  responses:
    Error400:
      description: Bad request, invalid input parameters
...
    Error403:
      description: Forbidden, the user does not have permission
    Error404:
      description: Requested resource not found
...
    Error500:
      description: Internal server error
    
#LOGIN UTENTE ATTRAVERSO USERNAME
paths:
...
  /conversation/{conversation_id}/names:
    parameters:
      - $ref: "#/components/parameters/ConversationId"
    put:
      tags: ["groups"]
      summary: Add users to a group
      description: "This endpoint adds one or more users to a specific group conversation based on the conversation ID"
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "name of the user to add in the group"
              properties:
                name:
                  $ref: "#/components/schemas/Name"
      responses:
        '204':
          description: "User(s) have been successfully added to the group"
        '404':
          $ref: "#/components/responses/Error404"
        '403':
          $ref: "#/components/responses/Error403"
        '400':
          $ref: "#/components/responses/Error400"
        '500':
          $ref: "#/components/responses/Error500"
      security:
        - bearerAuth: [] 
#ESCI DAL GRUPPO
...